common: retry if github commits URL fails
There is a race condition: sometimes github issues the webhook to us
and we try to retrieve the commits URL before github is ready.  Hence,
if we get a "message":"Not Found" response back in the JSON, then just
sleep a short period of time and try again.

Here's the JSON you'll get back in this case:

{
  "message":"Not Found",
  "documentation_url":"https://developer.github.com/v3/pulls/#list-commits-on-a-pull-request"
}

Signed-off-by: Jeff Squyres <jeff@squyres.com>